ai_services/requirements.txt
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
requests>=2.31.0